Convert custom field type on start after load
