+++ packages/core/content-type-builder/server/utils/attributes.js
@@ -67,15 +67,6 @@ const formatAttribute = (key, attribute) => {
     };
   }
 
-  if (attribute.type === 'customField') {
-    const customField = strapi.container.get('custom-fields').get(attribute.customField);
-
-    return {
-      ...attribute,
-      type: customField.type,
-    };
-  }
-
   return attribute;
 };
